perf: store one byte per square in the square list
Piece codes run 0..12, so the per square array does not need an int. This
takes it from 256 to 64 bytes and removes three cache lines from the
Board footprint that every move generation and SEE call touches.

+++ src/types.h
@@ -105,7 +105,7 @@ typedef struct {
 
   uint64_t piecesCounts; // "material key" - pieces left on the board
 
-  int squares[64];         // piece per square
+  uint8_t squares[64];     // piece per square
   BitBoard occupancies[3]; // 0 - white pieces, 1 - black pieces, 2 - both
   BitBoard pieces[12];     // individual piece data
